Update hello-world.md
Modified Netcat example to be GNU specific netcat command and added quote area with information about handling if user is using a different version of netcat. Added link to NMap.org version of netcat which is newer and has downloads for all major OS's, in case user needs to install.